server/requirements.txt
flask
google-generativeai
groq
python-dotenv
Pillow
flask-cors
requests
gunicorn
python-json-logger
